# docs/api-docs.yaml
openapi: 3.0.1
info:
  title: Tasa-API
  description: "API for synchronize TASA app data."
  version: "1.0"
servers:
  - url: http://localhost:8080/api
...
paths:
...
  /event/{id}:
    get:
      security:
        - bearerAuth: []
      tags:
        - Events
      summary: Retrieves a event
      description: Get a event given its id and respetive calendar id.
      parameters:
        - name: id
          in: path
          description: Event identifier
          required: true
          schema:
            type: integer
            format: int32
            example: 5287
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventOutput'
        '400':
          description: Event identifier is negative.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Event not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /event/all:
    get:
      security:
        - bearerAuth: []
      tags:
        - Events
      summary: Retrieves all events of a user
      description: Get all the registerd events of an user.
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventList'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
  /location/{id}:
    get:
      security:
        - bearerAuth: []
      tags:
        - Locations
      summary: Gets a location.
      description: Gets a location by it's identifier.
      parameters:
        - name: id
          in: path
          description: Location's identifier
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '400':
          description: Name is blank
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Location not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /location/all:
    get:
      security:
        - bearerAuth: []
      tags:
        - Locations
      summary: Gets all locations.
      description: Gets all locations of a user.
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationList'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
  /rule/location/{id}/update:
    put:
      security:
        - bearerAuth: []
      tags:
        - Rules
      summary: Updates a rule location
      description: Updates a rule location with a new start time and end time.
      parameters:
        - name: id
          in: path
          description: Rule identifier
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        description: Rule new end and start time.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RuleLocationUpdateInput"
        required: true
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleLocationOutput'
        '400':
          description: StartTime is blank
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Rule not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /rule/event/{id}/update/time:
    put:
      security:
        - bearerAuth: []
      tags:
        - Rules
      summary: Updates a rule event
      description: Updates a rule event new start time and end time.
      parameters:
        - name: id
          in: path
          description: Rule identifier
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        description: Rule data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RuleEventUpdateInput"
        required: true
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleEventOutput'
        '400':
          description: StartTime is blank
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Rule not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /rule/location:
    post:
      security:
        - bearerAuth: []
      tags:
        - Rules
      summary: Creates a rule location
      description: Creates a rule of silence for a location given a start time and end time.
      requestBody:
        description: Rule data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RuleLocationInput"
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleLocationOutput'
        '400':
          description: Start time is before end time.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: A rule already exists for given Location and time.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /rule/event:
    post:
      security:
        - bearerAuth: []
      tags:
        - Rules
      summary: Creates a rule event
      description: Creates a rule of silence for a event given a start time and end time.
      requestBody:
        description: Rule data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RuleEventInput"
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleEventOutput'
        '400':
          description: Start time is before end time.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: A rule already exists for given Event and time.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /rule/location/{id}:
    get:
      security:
        - bearerAuth: []
      tags:
        - Rules
      summary: Gets a rule location
      description: Gets a rule location by it's identifier.
      parameters:
        - name: id
          in: path
          description: rule identifier
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleLocationOutput'
        '400':
          description: Identifier is negative.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Rule not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

    delete:
      security:
        - bearerAuth: []
      tags:
        - Rules
      summary: Deletes a rule location
      description: Deletes a rule location by it's identifier.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Ok
        '400':
          description: Identifier is negative.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Rule not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /rule/event/{id}:
    get:
      security:
        - bearerAuth: []
      tags:
        - Rules
      summary: Gets a rule event
      description: Gets a rule event by it's identifier.
      parameters:
        - name: id
          in: path
          description: Rule identifier
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleEventOutput'
        '400':
          description: Identifier is negative.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Rule not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

    delete:
      security:
        - bearerAuth: []
      tags:
        - Rules
      summary: Deletes a rule event
      description: Deletes a rule event by it's identifier.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Ok
        '400':
          description: Identifier is negative.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Rule not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /rule/all:
    get:
      security:
        - bearerAuth: []
      tags:
        - Rules
      summary: Get's all rules.
      description: Get's all rules from a user.
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleListOutput'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    RuleLocationUpdateInput:
      type: object
      description: New data for rule.
      required:
        - endTime
        - startTime
      properties:
        startTime:
          type: string
          format: date-time
          example: "2024-10-01T12:00:00Z"
        endTime:
          type: string
          format: date-time
          example: "2024-10-01T12:00:00Z"
    RuleEventUpdateInput:
      type: object
      required:
        - endTime
        - startTime
      properties:
        startTime:
          type: string
          format: date-time
          example: "2024-10-01T12:00:00Z"
        endTime:
          type: string
          format: date-time
          example: "2024-10-01T12:00:00Z"
...
    EventOutput:
      type: object
      required:
        - id
        - title
        - endTime
        - startTime
      properties:
        id:
          type: integer
          format: int32
          description: Event identifier
          example: 1025
        startTime:
          type: string
          format: date-time
          example: "2024-10-01T12:00:00Z"
        endTime:
          type: string
          format: date-time
          example: "2024-10-01T12:00:00Z"
        title:
          type: string
          description: Event title
          example: Meeting.
    EventList:
      type: array
      properties:
        nEvents:
          type: integer
          description: number of results
          example: 20
        events:
          type: array
          items:
            $ref: '#/components/schemas/EventOutput'
...
    RuleLocationInput:
      required:
        - endTime
        - locationId
        - startTime
      type: object
      properties:
        startTime:
          type: string
          format: date-time
          description: Start time of rule
          example: "2024-10-01T12:00:00Z"
        endTime:
          type: string
          format: date-time
          description: End time of rule
          example: "2024-10-01T12:00:00Z"
        locationId:
          type: integer
          format: int32
          description: Location identifier.
          example: 10
    RuleEventInput:
      type: object
      required:
        - endTime
        - eventId
        - startTime
      properties:
        eventId:
          type: integer
          format: int32
          description: Event identifier.
          example: 1025
        startTime:
          type: string
          format: date-time
          description: Start time of rule
          example: "2024-10-01T12:00:00Z"
        endTime:
          type: string
          format: date-time
          description: End time of rule
          example: "2024-10-01T12:00:00Z"
...
    Location:
      type: object
      required:
        - id
        - latitude
        - longitude
        - name
        - radius
      properties:
        id:
          type: integer
          description: Location dentifier
          example: 10
        name:
          type: string
          description: Location name.
          example: "Home"
        latitude:
          type: number
          format: double
          description: Latitude of location.
          example: 38.75691186259142
        longitude:
          type: number
          format: double
          description: Langitude of location.
          example: -9.116494150002076
        radius:
          type: number
          format: double
          description: Radius of location.
          example: 30.0
    LocationList:
      type: object
      properties:
        nUsers:
          type: integer
          description: number of results
          example: 20
        users:
          type: array
          items:
            $ref: '#/components/schemas/Location'
    RuleLocationOutput:
      required:
        - id
        - endTime
        - locationId
        - startTime
      type: object
      properties:
        id:
          type: integer
          description: Rule dentifier
          example: 10
        startTime:
          type: string
          format: date-time
          description: Start time of rule
          example: "2024-10-01T12:00:00Z"
        endTime:
          type: string
          format: date-time
          description: End time of rule
          example: "2024-10-01T12:00:00Z"
        locationId:
          type: integer
          format: int32
          description: Location identifier.
          example: 10
    RuleEventOutput:
      type: object
      required:
        - id
        - endTime
        - startTime
        - event
      properties:
        id:
          type: integer
          description: Rule dentifier
          example: 10
        startTime:
          type: string
          format: date-time
          description: Start time of rule
          example: "2024-10-01T12:00:00Z"
        endTime:
          type: string
          format: date-time
          description: End time of rule
          example: "2024-10-01T12:00:00Z"
        event:
          $ref: '#/components/schemas/EventOutput'
    RuleEventOutputList:
      type: object
      properties:
        nUsers:
          type: integer
          description: number of results
          example: 20
        users:
          type: array
          items:
            $ref: '#/components/schemas/RuleEventOutput'
    RuleLocationOutputList:
      type: object
      properties:
        nUsers:
          type: integer
          description: number of results
          example: 20
        users:
          type: array
          items:
            $ref: '#/components/schemas/RuleLocationOutput'
    RuleListOutput:
      type: object
      properties:
        eventRulesN:
          type: integer
          description: number of results
          example: 20
        eventsRules:
          type: array
          items:
            $ref: '#/components/schemas/RuleEventOutputList'
        locationRulesN:
          type: integer
          description: number of results
          example: 20
        locationRules:
          type: array
          items:
            $ref: '#/components/schemas/RuleLocationOutputList'
    Problem:
      description: Error response
      type: object
      properties:
        type:
          type: string
        title:
          type: string
          format: uri
        status:
          type: integer
        detail:
          type: string